Add PCD support and organize encoding helpers
